Request data from the Flickr API

Request data from the Flickr API using the provided method, API key, and any additional values passed to httr2::req_url_query().

Usage

FlickrAPIRequest(
  method = NULL,
  api_key = NULL,
  format = "json",
  simplifyVector = TRUE,
  check_type = FALSE,
  ...,
  error_call = caller_env()
)

Arguments

method

Flickr API method to use for request.

api_key

Flickr API key. If api_key is NULL, the function uses getFlickrAPIKey() to use the environment variable "FLICKR_API_KEY" as the key.

format

Format parameter passed to httr2::req_url_query()

simplifyVector

Default to TRUE, passed to httr2::resp_body_json()

check_type

Default to FALSE, passed to httr2::resp_body_json()

...

Additional parameters passed to httr2::req_url_query()

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.