Get EXIF data for a Flickr photo.

Return a data of EXIF data for a given photo. The calling user must have permission to view the photo.

Usage

getExif(api_key = NULL, photo_id = NULL, secret = NULL)

get_exif(api_key = NULL, photo_id = NULL, secret = NULL)

Arguments

api_key

Flickr API key. If api_key is NULL, the function uses getFlickrAPIKey() to use the environment variable "FLICKR_API_KEY" as the key.

photo_id

The id of the photo to fetch EXIF information for.

secret

The secret for the photo. If the correct secret is passed then permissions checking is skipped. This enables the 'sharing' of individual photos by passing around the id and secret.

Value

A data.frame with EXIF information for an image with the provided photo_id.

See also

Examples

if (FALSE) { # \dontrun{
getExif(
  api_key = get_flickr_api_key(),
  photo_id = "45961963324"
)
} # }