Get a data frame of hot tags for a given time period.

Get a data frame of hot tags for a given time period.

Usage

getHotTags(api_key = NULL, period = c("day", "week"), count = 20)

get_hot_tags(api_key = NULL, period = c("day", "week"), count = 20)

Arguments

api_key

Flickr API key. If api_key is NULL, the function uses getFlickrAPIKey() to use the environment variable "FLICKR_API_KEY" as the key.

period

The period for which to fetch hot tags. Valid values are day or week. Defaults to day

count

The number of tags to return. Defaults to 20. Maximum allowed value is 200.

Value

This function a data frame of hot tags for the given period

See also

Examples

if (FALSE) { # \dontrun{
getHotTags(
  api_key = get_flickr_api_key(),
  period = "day",
  count = 20
)
} # }